Run classification in tl_detector thread
Move classification processing from image callback to the tl_detector
thread. This to make it run at lower priority such that it doesn't
prevent other parts of the system from running.

The synchronization is made using a mutex, and it's only keeping the
last image in the pipe. To make sure ROS doesn't mess around with the
image after returning the callback, the conversion to cv_image is
still implemented directly in the callback.